+++ libs/react/ui/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @shipfox/react-ui
 
+## 0.25.0
+
+### Minor Changes
+
+- e0e55ad: Update ModalBodyHeader and Table components
+- e0e55ad: Update ModalBodyHeader component
+
+## 0.24.0
+
+### Minor Changes
+
+- 7a9dad5: Add ModalBodyHeader component
+
 ## 0.23.0
 
 ### Minor Changes